+++ Sources/StreamVideoSwiftUI/Utils/PictureInPicture/PictureInPictureController.swift
@@ -65,6 +65,20 @@ final class PictureInPictureController: @unchecked Sendable {
             .sinkTask { @MainActor [weak self] in self?.didUpdate($0) }
             .store(in: disposableBag)
 
+        proxyDelegate
+            .publisher
+            .compactMap {
+                switch $0 {
+                case let .failedToStart(_, error):
+                    return error
+                default:
+                    return nil
+                }
+            }
+            .log(.error, subsystems: .pictureInPicture) { "Picture-in-Picture failed to start: \($0)." }
+            .sink { _ in }
+            .store(in: disposableBag)
+
         // Add delay to prevent premature cancellation
         applicationStateAdapter
             .statePublisher
@@ -78,8 +92,11 @@ final class PictureInPictureController: @unchecked Sendable {
     /// Updates the Picture-in-Picture controller when the source view changes.
     @MainActor
     private func didUpdate(_ sourceView: UIView?) {
-        guard let sourceView else {
+        guard let sourceView, store.state.call != nil else {
+            /// We ensure to cleanUp every Picture-in-Picture interacting component so that the next
+            /// Call will start with clean state.
             pictureInPictureController?.contentSource = nil
+            contentViewController = nil
             pictureInPictureController = nil
             disposableBag.remove(DisposableKey.isPossible.rawValue)
             disposableBag.remove(DisposableKey.isActive.rawValue)
@@ -96,16 +113,27 @@ final class PictureInPictureController: @unchecked Sendable {
         guard let contentViewController else {
             return
         }
-
-        pictureInPictureController = .init(
-            contentSource: .init(
-                activeVideoCallSourceView: sourceView,
-                contentViewController: contentViewController
-            )
+        let contentSource = AVPictureInPictureController.ContentSource(
+            activeVideoCallSourceView: sourceView,
+            contentViewController: contentViewController
         )
-        pictureInPictureController?.canStartPictureInPictureAutomaticallyFromInline = store.state
-            .canStartPictureInPictureAutomaticallyFromInline
-        pictureInPictureController?.delegate = proxyDelegate
+
+        if pictureInPictureController == nil {
+            pictureInPictureController = .init(
+                contentSource: contentSource
+            )
+            pictureInPictureController?.canStartPictureInPictureAutomaticallyFromInline = store
+                .state
+                .canStartPictureInPictureAutomaticallyFromInline
+            pictureInPictureController?.delegate = proxyDelegate
+        } else {
+            pictureInPictureController?.contentSource = contentSource
+        }
+
+        log.debug(
+            "SourceView updated and contentViewController preferredContentSize:\(contentViewController.preferredContentSize)",
+            subsystems: .pictureInPicture
+        )
     }
 
     /// Updates the content view controller when the view factory changes.